Replace original wgctrl
