+++ examples/full.conf
@@ -3,8 +3,7 @@
 [Interface]
 ListenPort = 51820
 PrivateKey = LEvpMepC4lqD7CJghor8f1cwA4E5r6AXrVRkXQQav6s=
-# FwMark is not supported
-# FwMark = 0
+FwMark = 10   # or FwMark = 0x11
 # PostUp and PostDown hooks are not supported.
 
 [Peer]
@@ -13,7 +12,6 @@ PresharedKey = 3D+23xivKLQntkTKXWf+0Zi8Nfq6Y6JooP9MoeGrV8U=
 AllowedIPs = 10.200.200.2/32
 Endpoint = www.example.com:880
 PersistentKeepalive = 30   # in seconds
-
 
 # Another peer
 # [Peer]
